refactor(ecstore): isolate metadata quorum decisions
