+++ crates/ecstore/src/set_disk/core/mod.rs
@@ -18,3 +18,4 @@
 //! duplicating read/write/erasure logic.
 
 pub(crate) mod io_primitives;
+mod metadata_quorum;
